HW1/hw1/src/main/java/tqs/homework/hw1/controllers/CityController.java
package tqs.homework.hw1.controllers;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import lombok.AllArgsConstructor;
import tqs.homework.hw1.models.City;
import tqs.homework.hw1.services.CityService;

import java.net.URI;

import org.springframework.http.ResponseEntity;

@RestController
@AllArgsConstructor
@RequestMapping("/cities")
public class CityController {

    private final CityService cityService;

    @GetMapping("/{id}")
    public ResponseEntity<City> getCityById(@PathVariable Long id) {
        City city = cityService.getCityById(id);
        if (city == null) {
            return ResponseEntity.notFound().build();
        }
        return ResponseEntity.ok(city);
    }

    @PostMapping
    public ResponseEntity<City> saveCity(@RequestBody City city) {
        City savedCity = cityService.saveCity(city);
        URI location = ServletUriComponentsBuilder
                .fromCurrentRequest()
                .path("/{id}")
                .buildAndExpand(savedCity.getId())
                .toUri();
        return ResponseEntity.created(location).body(savedCity);
    }

    @GetMapping
    public ResponseEntity<Iterable<City>> getAllCities() {
        return ResponseEntity.ok(cityService.getAllCities());
    }
}
